pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "livigy-ups-ha-bridge"
version = "0.1.0"
description = "Livigy/PowerShield UPS to Home Assistant MQTT bridge"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "paho-mqtt==2.1.0",
  "PyYAML==6.0.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
